Correction bug foreign keys.

+++ database/migrations/2017_04_06_144249_create_partners_menus_table.php
@@ -25,7 +25,7 @@ class CreatePartnersMenusTable extends Migration
                 ->onDelete('cascade')
                 ->onUpdate('cascade');
 
-            $table->bigInteger('category_id')->unsigned();
+            $table->integer('category_id')->unsigned();
             $table->foreign('category_id')
                 ->references('id')
                 ->on('menu_categories');
